board 도메인에 맞는 컨트롤러 제작 및 테스트 작성
